Netlink: Don't rely on notifications
When creating a veth pair we were relying on rtnetlink notifications
as a way of getting back the created devices. This is not a good idea
since notifications are system wide and there are no guarantees around
ordering. With this patch, we explicitly issue a get request to the
kernel.

Additional fixes were applied to the rtnetlink library.

Change-Id: Ieb8960eda5ed181130b14c2b5a9ef1b1d44673c3
Signed-off-by: Duarte Nunes <duarte@midokura.com>